# server/api/openapi.yaml
openapi: "3.0.3"
info:
  title: "Fast Ship API"
  version: "0.1.42"
  license:
    name: MIT
  description: |
    Fast Ship 项目管理工具的 HTTP API。覆盖 `server/internal/router/router.go` 注册的全部 /api 路由；
    SPA 静态资源与 NoRoute 回退（`web.go`）不在本文档范围。

    约定：
    - 所有 JSON 响应统一信封 `{code, message, data}`：`code=0` 成功；非 0 为业务错误码，`data` 为 null。
    - 认证方式为 `Authorization: Bearer <token>`；token 为 JWT（登录/注册签发）或 `fsk_` 开头的 API Key，服务端按前缀区分。
    - 部分二进制资源端点（头像、Issue 图片、安装包下载、GitHub 媒体代理）额外接受 `?token=<token>` query 凭证。
    - 除特别注明外，字段级权限写在各字段 description 中：「仅 JWT」表示 API Key 携带该字段会返回 403（40301）；「仅 API Key」表示 JWT 调用返回 403（40303）。
servers:
  - url: /
    description: 与后端服务同源
...
paths:
...
  /api/issues/assets/{aid}/content:
    get:
      tags: [issues]
      operationId: getIssueAssetContent
      summary: 获取 Issue 图片附件内容
      description: "返回附件二进制内容（实际 mime type，带 `Cache-Control: private, max-age=300`）。草稿附件与已挂载附件均可。同一路径另注册 HEAD 变体，语义同 GET 但不返回 body。"
      security:
        - jwtAuth: []
        - apiKeyAuth: []
      parameters:
        - name: aid
          in: path
          required: true
          schema:
            type: string
          description: 附件 ID（IssueAsset 或 IssueDraftAsset 的 UUID）
        - $ref: '#/components/parameters/QueryToken'
      responses:
        '200':
          description: 附件二进制内容
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    head:
      tags: [issues]
      operationId: getIssueAssetContentHead
      summary: 获取 Issue 图片附件头信息（HEAD）
      description: 与 GET 语义相同但不返回 body。
      security:
        - jwtAuth: []
        - apiKeyAuth: []
      parameters:
        - name: aid
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/QueryToken'
      responses:
        '200':
          description: 仅响应头
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/projects/{id}/issues/count:
    get:
      tags: [issues]
      operationId: countIssues
      summary: 按过滤条件统计 Issue 数
      security:
        - jwtAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: state
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/IssueState'
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: 同 listIssues 的 q：标题/正文/作者/编号/纯数字序号
        - name: label
          in: query
          required: false
          schema:
            type: string
        - name: source
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/IssueSource'
        - name: assignee
          in: query
          required: false
          schema:
            type: string
        - name: milestone
          in: query
          required: false
          schema:
            type: string
        - name: workflow_status
          in: query
          required: false
          schema:
            type: string
            enum: [todo, in_progress, done, unset]
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: 接受但忽略（计数不排序）
      responses:
        '200':
          description: 匹配数量
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [count]
                        properties:
                          count:
                            type: integer
                            format: int64
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/projects/{id}/issues/batch-close:
    post:
      tags: [issues]
      operationId: batchCloseDoneIssues
      summary: 批量关闭 workflow_status=done 的 open Issue（仅 JWT；超过 200 条返回 41201）
      security:
        - jwtAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                source:
                  $ref: '#/components/schemas/IssueSource'
                  description: 可选，限定仅关闭 internal 或 github 来源；缺省全部
      responses:
        '200':
          description: 批量关闭结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IssueBatchResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          description: 匹配问题数量超过 200（41201）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/issues/{iid}/ship-hook:
    put:
      tags: [issues]
      operationId: upsertShipHook
      summary: 预约发货后钩子（仅 JWT；整体覆盖）
      description: 配置该 Issue 所属项目下一次任意版本成功发货后执行的一次性动作；三类动作至少启用一个。comment_body 中 `{version}`、`{release_url}` 占位符在发货时替换。
      security:
        - jwtAuth: []
      parameters:
        - $ref: '#/components/parameters/IssueId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertShipHookRequest'
      responses:
        '200':
          description: 更新后的钩子
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IssueShipHook'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [issues]
      operationId: deleteShipHook
      summary: 删除发货后钩子（仅 JWT）
      security:
        - jwtAuth: []
      parameters:
        - $ref: '#/components/parameters/IssueId'
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/issues/{iid}/read:
    post:
      tags: [issues]
      operationId: markIssueRead
      summary: 标记 Issue 评论已读（仅 JWT；internal Issue 无未读概念，恒成功）
      security:
        - jwtAuth: []
      parameters:
        - $ref: '#/components/parameters/IssueId'
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token，由 POST /api/auth/register、login 或 refresh 签发。仅 JWT 的端点对 `fsk_` API Key 返回 403（40301）。
    apiKeyAuth:
      type: http
      scheme: bearer
      description: "API Key，以 `fsk_` 开头（Web「设置 → API Key」创建）。与 jwtAuth 共用 `Authorization: Bearer` 头，服务端按 `fsk_` 前缀区分凭证类型。"
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 项目 ID（UUID）
...
    IssueId:
      name: iid
      in: path
      required: true
      schema:
        type: string
      description: Issue ID（UUID，非 INT-123/GH-123 短编号）
...
    QueryToken:
      name: token
      in: query
      required: false
      schema:
        type: string
      description: 可选 query 凭证（JWT 或 `fsk_` API Key），供无法携带 Authorization 头的场景（`<img>`、浏览器直链下载）。与 Authorization 头二选一，头优先。
  responses:
    EmptySuccess:
      description: 操作成功（data 为 null）
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data:
                    type: object
                    nullable: true
    BadRequest:
      description: 请求参数无效（40001-40099）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: 未提供或提供无效凭证（40100-40199）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: 已认证但无权限（40300-40399；40301=API Key 越权，40303=仅限 API Key）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: 资源不存在（40400-40499）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    InternalError:
      description: 服务器内部错误（50000）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    Envelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 业务错误码；0 表示成功
        message:
          type: string
          description: 人类可读信息；成功时为 "success"
        data:
          description: 响应负载；形状随端点而变，由 allOf 组合的兄弟 schema 收窄
      description: 统一 JSON 信封 {code, message, data}
    ErrorEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          description: 业务错误码（4xxxx 按区间映射 HTTP 状态：401xx→401，403xx→403，404xx→404，409xx→409，412xx→412，50xxx→500；502xx→502）
        message:
          type: string
        data:
          type: object
          nullable: true
          description: 恒为 null
      description: 错误响应信封
...
    IssueState:
      type: string
      enum: [open, closed]
      x-go-type: model.IssueState
      x-go-type-import:
        name: model
        path: github.com/godbobo/fast_ship/server/internal/model
    IssueSource:
      type: string
      enum: [github, internal]
      x-go-type: model.IssueSource
      x-go-type-import:
        name: model
        path: github.com/godbobo/fast_ship/server/internal/model
    IssueWorkflowStatus:
      type: string
      enum: ["", todo, in_progress, done]
      description: 空串表示未设置（重置语义）
      x-go-type: model.IssueWorkflowStatus
      x-go-type-import:
        name: model
        path: github.com/godbobo/fast_ship/server/internal/model
...
    IssueShipHookStatus:
      type: string
      enum: [pending, running, fired]
      x-go-type: model.IssueShipHookStatus
      x-go-type-import:
        name: model
        path: github.com/godbobo/fast_ship/server/internal/model

    # ---- auth ----
...
    UpsertShipHookRequest:
      type: object
      description: 三类动作至少启用一个，否则 40001
      properties:
        comment_body:
          type: string
          description: 传入即启用评论动作；trim 后 1-4000 rune；支持 {version}、{release_url} 占位符
        close:
          type: boolean
          description: true 时发货后关闭 Issue（state_reason=completed）
        workflow_status:
          $ref: '#/components/schemas/IssueWorkflowStatus'
          description: 传入即启用工作流动作（含 "" 重置）
...
    IssueBatchFailure:
      type: object
      required: [id, error]
      properties:
        id:
          type: string
          description: Issue ID
        reference:
          type: string
          description: omitempty；INT-n 或 GH-n
        error:
          type: string
    IssueBatchResult:
      type: object
      required: [total, succeeded, failed, failures, elapsed_ms]
      properties:
        total:
          type: integer
          format: int64
        succeeded:
          type: integer
        failed:
          type: integer
        failures:
          type: array
          items:
            $ref: '#/components/schemas/IssueBatchFailure'
          description: 最多返回前 50 条失败明细
        elapsed_ms:
          type: integer
          format: int64
      description: 批量操作结果，与 service.BatchCloseDoneIssuesResponse 同形；batch-close 与批量 internal-meta 共用
...
    IssueShipHookActionResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        skipped:
          type: boolean
          description: omitempty
        error:
          type: string
          description: omitempty
    IssueShipHookResults:
      type: object
      properties:
        comment:
          $ref: '#/components/schemas/IssueShipHookActionResult'
        close:
          $ref: '#/components/schemas/IssueShipHookActionResult'
        workflow_status:
          $ref: '#/components/schemas/IssueShipHookActionResult'
      description: omitempty；fired 后按启用的动作出现
    IssueShipHook:
      type: object
      required: [status, comment_enabled, close_enabled, workflow_enabled, workflow_status]
      properties:
        status:
          $ref: '#/components/schemas/IssueShipHookStatus'
        comment_enabled:
          type: boolean
        comment_body:
          type: string
          description: omitempty；fired 后为渲染后正文
        close_enabled:
          type: boolean
        workflow_enabled:
          type: boolean
        workflow_status:
          type: string
          description: 目标工作流状态；workflow_enabled=true 且为 "" 表示重置
        version_id:
          type: string
          description: omitempty；fired 时的触发版本
        version_number:
          type: string
          description: omitempty
        release_url:
          type: string
          description: omitempty
        fired_at:
          type: string
          format: date-time
          description: omitempty
        results:
          $ref: '#/components/schemas/IssueShipHookResults'
      description: omitempty；仅 Issue 详情/列表项中出现，未预约时缺省
...